Refactor: Move MQTT connection to BridgeContext
